+++ query/client.go
@@ -10,6 +10,8 @@ import (
 )
 
 type Client interface {
+	Executor
+
 	// Do provide the best effort for execute operation.
 	//
 	// Do implements internal busy loop until one of the following conditions is met:
@@ -33,35 +35,19 @@ type Client interface {
 	// Warning: if context without deadline or cancellation func than DoTx can run indefinitely
 	DoTx(ctx context.Context, op TxOperation, opts ...DoTxOption) error
 
-	// Exec is a simple executor with retries
-	//
-	// Execute executes query without result
-	//
-	// Warning: large result can lead to "OOM Killed" problem
-	//
-	// Experimental: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#experimental
-	Exec(ctx context.Context, query string, opts ...options.ExecOption) error
-
-	// Query is a simple executor with retries which returns materialized result
-	//
-	// Warning: large result can lead to "OOM Killed" problem
-	//
-	// Experimental: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#experimental
-	Query(ctx context.Context, query string, opts ...options.QueryOption) (Result, error)
-
 	// QueryResultSet is a helper which read all rows from first result set in result
 	//
 	// ReadRow returns error if result contains more than one result set
 	//
 	// Experimental: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#experimental
-	QueryResultSet(ctx context.Context, query string, opts ...options.QueryOption) (ResultSet, error)
+	QueryResultSet(ctx context.Context, query string, opts ...options.Execute) (ResultSet, error)
 
 	// QueryRow is a helper which read only one row from first result set in result
 	//
 	// ReadRow returns error if result contains more than one result set or more than one row
 	//
 	// Experimental: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#experimental
-	QueryRow(ctx context.Context, query string, opts ...options.QueryOption) (Row, error)
+	QueryRow(ctx context.Context, query string, opts ...options.Execute) (Row, error)
 }
 
 type (